refactor(settings_ui): lift advanced-settings handlers to _handlers_advanced.py
Decomposes _save_advanced_settings (C901 33 -> under) via
_validate_advanced_batch/_validate_one_advanced_field/_coerce_advanced_value/
_check_advanced_bounds_choices/_advanced_env_pin_rejection + _save_advanced_addon/
_write_advanced_overrides_file, and factors _get_advanced_settings row-building into
_advanced_field_row. Last handler family — build_settings_handlers is now a thin assembler.
