+++ src/ha_mcp/settings_ui/__init__.py
@@ -34,7 +34,7 @@ from ..config import (
 from ..errors import ErrorCode, create_error_response
 from ..transforms import DEFAULT_PINNED_TOOLS
 from ..utils.data_paths import get_data_dir
-from . import _handlers_server, _persistence, _supervisor
+from ._handlers_advanced import build_advanced_handlers
 from ._handlers_backups import build_backups_handlers
 from ._handlers_fs import build_fs_handlers
 from ._handlers_server import (
@@ -377,510 +377,9 @@ def build_settings_handlers(
     async def _settings_page(_: Request) -> HTMLResponse:
         return HTMLResponse(_render_settings_html())
 
-    async def _get_advanced_settings(_: Request) -> JSONResponse:
-        """Return advanced (non-feature-flag, non-backup) settings + per-field
-        origin + editable flag.
-
-        Mirrors ``_get_feature_flags`` / ``_get_backup_config`` but for
-        the ``ADVANCED_SETTINGS_FIELDS`` registry. Most advanced fields
-        write to ``feature_flags.json`` via the shared override file in
-        either deployment mode. ``ADDON_SYNCED_ADVANCED_FIELDS``
-        (currently ``backup_hint``, ``verify_ssl``) are an exception:
-        in addon mode they have ``origin="addon"`` (editable) and saves
-        route through Supervisor ``/addons/self/options`` so the addon
-        Configuration tab and the web UI share state.
-        """
-        from ..config import (
-            _ADVANCED_SETTINGS_BOUNDS,
-            _ADVANCED_SETTINGS_CHOICES,
-            _ADVANCED_SETTINGS_SENTINELS,
-            ADVANCED_SETTINGS_FIELDS,
-            OAUTH_MODE_TOKEN,
-            _read_feature_flag_override_file,
-        )
-
-        settings = get_global_settings()
-        # Read the override file ONCE for this GET — origin lookup is
-        # called per field, and re-reading the file 17+ times would
-        # produce duplicate WARNINGs on a corrupt file (one per field).
-        overrides = _read_feature_flag_override_file()
-        # Read-consistency with the add-on Configuration tab: addon-synced
-        # fields (origin "addon") live in /data/options.json, so the boot-env
-        # value goes stale after a config edit without a restart. Surface the
-        # latest SAVED options value for those (see _live_addon_options).
-        live_options = await _handlers_server._live_addon_options(server)
-        fields: list[dict[str, Any]] = []
-        for (
-            fname,
-            env_name,
-            ftype,
-            section,
-            registry_editable,
-        ) in ADVANCED_SETTINGS_FIELDS:
-            origin = _origin_for_advanced_field(env_name, overrides=overrides)
-            value: Any = getattr(settings, fname, None)
-            if origin == "addon" and fname in live_options:
-                value = live_options[fname]
-            # Mask the token: never echo the actual long-lived access
-            # token to the UI. The OAuth-mode sentinel survives so
-            # operators can tell connection mode at a glance.
-            if fname == "homeassistant_token":
-                value = "*****" if value and value != OAUTH_MODE_TOKEN else value
-            row: dict[str, Any] = {
-                "field": fname,
-                "env_var": env_name,
-                "value": value,
-                "type": ftype.__name__,
-                "section": section,
-                "origin": origin,
-                # Env-pin makes the field read-only regardless of the
-                # registry's ``editable`` flag. Display-only rows from
-                # the registry (homeassistant_url / _token) stay locked
-                # forever.
-                "editable": registry_editable and origin != "env",
-            }
-            bounds = _ADVANCED_SETTINGS_BOUNDS.get(fname)
-            if bounds is not None:
-                row["min"], row["max"] = bounds
-                # Sentinel fields (e.g. sidecar_pin_port: 0 = off) need the
-                # number input to reach below the bounded range, so expose
-                # the sentinel as the UI minimum.
-                sentinel = _ADVANCED_SETTINGS_SENTINELS.get(fname)
-                if sentinel is not None:
-                    row["min"] = sentinel
-            choices = _ADVANCED_SETTINGS_CHOICES.get(fname)
-            if choices is not None:
-                row["choices"] = list(choices)
-            fields.append(row)
-        # is_stdio: the sidecar-port field only applies when this settings
-        # page is served by the stdio settings-UI sidecar. In HTTP/SSE/OAuth
-        # /addon deployments there is no sidecar, so the UI greys the section.
-        return JSONResponse(
-            {
-                "fields": fields,
-                "is_addon": is_running_in_addon(),
-                "is_stdio": get_http_settings_prefix() is None,
-            }
-        )
-
-    def _origin_for_advanced_field(
-        env_name: str, overrides: dict[str, Any] | None = None
-    ) -> str:
-        """Origin for an ADVANCED_SETTINGS_FIELDS entry.
-
-        Returns ``'addon' | 'env' | 'file' | 'default'``.
-
-        ``'addon'`` is returned in addon mode for fields that live in
-        both the registry and the addon's config.yaml schema (the
-        ``ADDON_SYNCED_ADVANCED_FIELDS`` set). For those, writes route
-        through Supervisor instead of the override file so the addon
-        Configuration tab and this web UI share state. Other
-        env-pinned fields stay ``'env'`` (locked).
-
-        Callers iterating ADVANCED_SETTINGS_FIELDS should pass a
-        pre-read ``overrides`` dict so the override file isn't re-read
-        N times per page render.
-        """
-        from ..config import (
-            ADDON_SYNCED_ADVANCED_FIELDS,
-            ADVANCED_SETTINGS_FIELDS,
-            _read_feature_flag_override_file,
-        )
-
-        fname = next(
-            (f for f, e, *_ in ADVANCED_SETTINGS_FIELDS if e == env_name), None
-        )
-        if (
-            is_running_in_addon()
-            and fname is not None
-            and fname in ADDON_SYNCED_ADVANCED_FIELDS
-        ):
-            return "addon"
-        if os.environ.get(env_name) is not None:
-            return "env"
-        if overrides is None:
-            overrides = _read_feature_flag_override_file()
-        if fname is not None and fname in overrides:
-            return "file"
-        return "default"
-
-    async def _save_advanced_settings(request: Request) -> JSONResponse:
-        """Persist UI-edited advanced settings.
-
-        Two persistence sinks depending on field origin:
-
-        - ``ADDON_SYNCED_ADVANCED_FIELDS`` (``backup_hint``,
-          ``verify_ssl``) in addon mode → write goes through
-          Supervisor ``/addons/self/options`` so the addon
-          Configuration tab and this web UI share state.
-        - Everything else → atomic write to ``feature_flags.json``
-          (the shared override file used by feature flags), then
-          ``_reset_global_settings()`` so the next read picks up the
-          new value.
-
-        Validation chain per field:
-        - Unknown field → 400 ``VALIDATION_INVALID_PARAMETER``.
-        - ``editable=False`` registry entry → 409 (display-only).
-        - Env-pinned (``origin=='env'`` and NOT addon-synced) → 409
-          with env var name.
-        - Type mismatch → 400.
-        - Bounds violation → 400.
-        - Choices violation → 400.
-
-        Either sink responds with ``restart_required=True`` so the UI
-        shows the banner — most advanced fields gate one-time startup
-        paths (REST client construction, logging setup, MCP handshake
-        metadata, tool-module filtering, etc.).
-        """
-        from ..config import (
-            _ADVANCED_SETTINGS_BOUNDS,
-            _ADVANCED_SETTINGS_CHOICES,
-            _ADVANCED_SETTINGS_SENTINELS,
-            _FEATURE_FLAG_OVERRIDE_FILENAME,
-            ADVANCED_SETTINGS_FIELDS,
-        )
-        from ..utils.data_paths import get_data_dir
-
-        try:
-            body = await request.json()
-        except (ValueError, TypeError):
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.VALIDATION_INVALID_JSON,
-                    "Invalid JSON body",
-                ),
-                status_code=400,
-            )
-        if not isinstance(body, dict):
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.VALIDATION_INVALID_PARAMETER,
-                    "Request body must be a JSON object",
-                ),
-                status_code=400,
-            )
-
-        from ..config import ADDON_SYNCED_ADVANCED_FIELDS
-
-        registry = {f: (e, t, s, ed) for f, e, t, s, ed in ADVANCED_SETTINGS_FIELDS}
-        new_overrides: dict[str, Any] = {}
-        addon_writes_present = False
-        addon_mode = is_running_in_addon()
-        for fname, raw in body.items():
-            if fname not in registry:
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.VALIDATION_INVALID_PARAMETER,
-                        f"Unknown advanced field: {fname!r}",
-                    ),
-                    status_code=400,
-                )
-            env_name, ftype, _section, registry_editable = registry[fname]
-            if not registry_editable:
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.VALIDATION_INVALID_PARAMETER,
-                        f"{fname!r} is display-only. Modify via env var "
-                        "or App (add-on) configuration.",
-                    ),
-                    status_code=409,
-                )
-            # Addon-synced fields (e.g. backup_hint, verify_ssl) are
-            # editable in addon mode even though their env vars are
-            # set — start.py rewrites them from /data/options.json on
-            # every boot, so we route the user's write through
-            # Supervisor instead of the override file.
-            is_addon_synced = addon_mode and fname in ADDON_SYNCED_ADVANCED_FIELDS
-            if is_addon_synced:
-                addon_writes_present = True
-            elif os.environ.get(env_name) is not None:
-                # Add-on mode has no env-var surface for non-schema keys
-                # (e.g. CODE_MODE_SAVED_TOOLS_PATH, set by start.py), so
-                # "unset it to edit here" is unactionable there — give
-                # add-on-aware copy instead of implying a lever exists.
-                if addon_mode:
-                    message = (
-                        f"{fname!r} is fixed by the App (add-on) runtime and "
-                        "cannot be changed from the web UI."
-                    )
-                    suggestions = [
-                        "This value is baked into the App (add-on) and is not "
-                        "exposed as an editable setting.",
-                    ]
-                else:
-                    message = (
-                        f"{fname!r} is set via {env_name} env var. "
-                        "Unset it to edit here."
-                    )
-                    suggestions = [
-                        f"Unset the {env_name} environment variable (or "
-                        "remove it from your Docker config), then restart "
-                        "to edit this setting from the UI.",
-                    ]
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.VALIDATION_INVALID_PARAMETER,
-                        message,
-                        suggestions=suggestions,
-                        context={"env_var": env_name},
-                    ),
-                    status_code=409,
-                )
-
-            coerced: Any
-            if ftype is bool:
-                if not isinstance(raw, bool):
-                    return _bad_advanced_type(fname, ftype, raw)
-                coerced = raw
-            elif ftype is int:
-                if isinstance(raw, bool) or not isinstance(raw, int):
-                    return _bad_advanced_type(fname, ftype, raw)
-                coerced = int(raw)
-            elif ftype is float:
-                if isinstance(raw, bool) or not isinstance(raw, int | float):
-                    return _bad_advanced_type(fname, ftype, raw)
-                coerced = float(raw)
-            elif ftype is str:
-                if not isinstance(raw, str):
-                    return _bad_advanced_type(fname, ftype, raw)
-                if "\x00" in raw:
-                    return JSONResponse(
-                        create_error_response(
-                            ErrorCode.VALIDATION_INVALID_PARAMETER,
-                            f"{fname!r} contains a null byte; rejected.",
-                        ),
-                        status_code=400,
-                    )
-                coerced = raw
-            else:
-                return _bad_advanced_type(fname, ftype, raw)
-
-            bounds = _ADVANCED_SETTINGS_BOUNDS.get(fname)
-            sentinel = _ADVANCED_SETTINGS_SENTINELS.get(fname)
-            if (
-                bounds is not None
-                and coerced != sentinel
-                and not (bounds[0] <= coerced <= bounds[1])
-            ):
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.VALIDATION_INVALID_PARAMETER,
-                        f"{fname!r} must be between {bounds[0]} and "
-                        f"{bounds[1]} (got {coerced}).",
-                        suggestions=[
-                            f"Provide a value for {fname} between "
-                            f"{bounds[0]} and {bounds[1]}.",
-                        ],
-                    ),
-                    status_code=400,
-                )
-            choices = _ADVANCED_SETTINGS_CHOICES.get(fname)
-            if choices is not None and coerced not in choices:
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.VALIDATION_INVALID_PARAMETER,
-                        f"{fname!r} must be one of {list(choices)} (got {coerced!r}).",
-                        suggestions=[
-                            f"Set {fname} to one of: {', '.join(map(str, choices))}.",
-                        ],
-                    ),
-                    status_code=400,
-                )
-            new_overrides[fname] = coerced
-
-        if not new_overrides:
-            return JSONResponse(
-                {
-                    "success": True,
-                    "applied": {},
-                    "mode": "file",
-                    "restart_required": False,
-                }
-            )
-
-        # Addon-route: at least one field is an addon-synced advanced
-        # field (backup_hint / verify_ssl in addon mode). Batch every
-        # write in this call into a single Supervisor options POST
-        # — same merge-then-replace pattern as the feature-flag
-        # addon-route, including the "single persistence path"
-        # invariant: we don't mix override-file writes and Supervisor
-        # writes from the same call. If a future caller submits both
-        # addon-synced and non-addon fields in one batch (none today
-        # are non-addon AND non-locked in addon mode), reject loudly
-        # instead of routing them through different sinks.
-        if addon_writes_present:
-            if not addon_mode:
-                # Defensive: addon_writes_present should imply addon_mode
-                # because is_addon_synced is gated on it above.
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.INTERNAL_ERROR,
-                        "Inconsistent addon-route classification",
-                    ),
-                    status_code=500,
-                )
-            file_only = {
-                k: v
-                for k, v in new_overrides.items()
-                if k not in ADDON_SYNCED_ADVANCED_FIELDS
-            }
-            if file_only:
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.INTERNAL_ERROR,
-                        (
-                            "Mixed addon-synced and override-file advanced "
-                            "writes in one batch; the UI should split these "
-                            f"into separate POSTs ({sorted(file_only)})."
-                        ),
-                        suggestions=[
-                            "Submit addon-synced fields (e.g. backup_hint, "
-                            "verify_ssl) and override-file fields in separate "
-                            "save requests.",
-                        ],
-                    ),
-                    status_code=500,
-                )
-            if server is None:
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.INTERNAL_ERROR,
-                        "Advanced settings POST requires a live MCP server",
-                    ),
-                    status_code=500,
-                )
-            ok, sup_err = await _supervisor._supervisor_merge_and_post_options(
-                server.settings.verify_ssl, new_overrides
-            )
-            if not ok:
-                if sup_err is None:
-                    # ``ok=False`` with no error is a contract bug in
-                    # the helper, not a user-actionable failure. Bail
-                    # with INTERNAL_ERROR instead of letting an
-                    # AttributeError leak under ``python -O`` (where
-                    # the previous ``assert`` was stripped).
-                    return JSONResponse(
-                        create_error_response(
-                            ErrorCode.INTERNAL_ERROR,
-                            "Supervisor helper returned ok=False with no error",
-                            suggestions=[
-                                "Check the Home Assistant Supervisor logs and "
-                                + "the App (add-on) logs for the underlying failure.",
-                                "Report this at "
-                                + "https://github.com/homeassistant-ai/ha-mcp/issues "
-                                + "if it persists. This indicates an internal bug.",
-                            ],
-                        ),
-                        status_code=500,
-                    )
-                # Mirror the sibling ``_save_feature_flags`` /
-                # ``_save_backup_config`` handlers: log loudly before
-                # returning so addon-log forensics survive the user
-                # closing the tab.
-                logger.warning(
-                    "Supervisor advanced-settings update failed (%s): %s",
-                    sup_err.kind,
-                    sup_err.message,
-                )
-                code = (
-                    ErrorCode.CONNECTION_FAILED
-                    if sup_err.kind == "transport"
-                    else ErrorCode.CONFIG_VALIDATION_FAILED
-                )
-                return JSONResponse(
-                    create_error_response(code, sup_err.message),
-                    status_code=sup_err.status_code,
-                )
-            return JSONResponse(
-                {
-                    "success": True,
-                    "applied": new_overrides,
-                    "mode": "addon",
-                    "restart_required": True,
-                }
-            )
-
-        # Merge into the existing override file via atomic write (same
-        # path used by _save_feature_flags so a single file holds both
-        # advanced + feature-flag overrides). Lock-serialised against
-        # concurrent feature-flag saves on the same file.
-        path = get_data_dir() / _FEATURE_FLAG_OVERRIDE_FILENAME
-        async with _persistence._get_override_file_lock():
-            existing: dict[str, Any] = {}
-            try:
-                existing_raw = path.read_text()
-            except FileNotFoundError:
-                existing_raw = None
-            except OSError as exc:
-                logger.warning("Cannot read %s", path, exc_info=True)
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.INTERNAL_ERROR,
-                        f"Could not read existing override file "
-                        f"({type(exc).__name__}: {exc}); refusing to "
-                        "overwrite to preserve prior toggles.",
-                    ),
-                    status_code=500,
-                )
-            if existing_raw is not None:
-                try:
-                    parsed = json.loads(existing_raw)
-                except json.JSONDecodeError as exc:
-                    logger.warning(
-                        "Existing %s is corrupt: %s", path, exc, exc_info=True
-                    )
-                    return JSONResponse(
-                        create_error_response(
-                            ErrorCode.VALIDATION_INVALID_PARAMETER,
-                            f"Existing override file at {path} is not "
-                            f"valid JSON ({exc}); refusing to overwrite. "
-                            "Inspect or delete the file manually and retry.",
-                        ),
-                        status_code=409,
-                    )
-                if isinstance(parsed, dict):
-                    existing = parsed
-            existing.update(new_overrides)
-
-            try:
-                _persistence._atomic_write_json(path, existing)
-            except OSError as exc:
-                logger.warning("Could not write %s", path, exc_info=True)
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.INTERNAL_ERROR,
-                        f"Could not persist advanced settings: {exc}",
-                    ),
-                    status_code=500,
-                )
-
-        _reset_global_settings()
-        return JSONResponse(
-            {
-                "success": True,
-                "applied": new_overrides,
-                "mode": "file",
-                "restart_required": True,
-            }
-        )
-
-    def _bad_advanced_type(fname: str, ftype: type, raw: Any) -> JSONResponse:
-        return JSONResponse(
-            create_error_response(
-                ErrorCode.VALIDATION_INVALID_PARAMETER,
-                f"{fname!r} expects {ftype.__name__}, got {type(raw).__name__}.",
-                suggestions=[
-                    f"Send {fname} as a {ftype.__name__} value.",
-                ],
-            ),
-            status_code=400,
-        )
-
     handlers: dict[str, Any] = {
         "root_page": _root_page,
         "settings_page": _settings_page,
-        "get_advanced_settings": _get_advanced_settings,
-        "save_advanced_settings": _save_advanced_settings,
     }
 
     # Tool security policies. The main server attaches an
@@ -911,6 +410,7 @@ def build_settings_handlers(
     handlers.update(build_tools_handlers(server))
     handlers.update(build_backups_handlers(server))
     handlers.update(build_server_handlers(server, is_sidecar=is_sidecar))
+    handlers.update(build_advanced_handlers(server))
 
     return handlers
